+++ lib/sidekiq/tasks/strategies/rake_task.rb
@@ -2,6 +2,11 @@ module Sidekiq
   module Tasks
     module Strategies
       class RakeTask < Base
+        def initialize(rules: [], scanner: Sidekiq::Tasks::MagicComments::Scanner.new)
+          super(rules: rules)
+          @scanner = scanner
+        end
+
         def load_tasks
           Rake::TaskManager.record_task_metadata = true
           Rake.application.load_rakefile
@@ -13,12 +18,22 @@ module Sidekiq
             name: task.name,
             desc: task.full_comment,
             file: task.locations.first.split(":").first,
-            args: task.arg_names
+            args: task.arg_names,
+            sidekiq_options: extract_sidekiq_options(task)
           )
         end
 
         def execute_task(name, args = nil)
           Rake::Task[name].execute(args)
+        end
+
+        private
+
+        def extract_sidekiq_options(task)
+          @scanner.scan(task).fetch(
+            Sidekiq::Tasks::MagicComments::Handlers::SidekiqOptions.name_token,
+            default: {}
+          )
         end
       end
     end