+++ lib/sidekiq/tasks/task_metadata.rb
@@ -3,13 +3,14 @@ module Sidekiq
     class TaskMetadata
       include Sidekiq::Tasks::Validations
 
-      attr_reader :name, :desc, :file, :args
+      attr_reader :name, :desc, :file, :args, :sidekiq_options
 
-      def initialize(name:, file:, desc: "", args: [])
+      def initialize(name:, file:, desc: "", args: [], sidekiq_options: {})
         @name = name
         @file = file
         @desc = desc
         @args = args
+        @sidekiq_options = sidekiq_options
 
         validate_params!
       end
@@ -21,6 +22,7 @@ module Sidekiq
         validate_class!(file, [String, NilClass], "file")
         validate_class!(desc, [String, NilClass], "desc")
         validate_array_classes!(args, [String, Symbol], "args")
+        Sidekiq::Tasks::SidekiqOptionsValidator.validate!(sidekiq_options, allow_partial: true)
       end
     end
   end