+++ spec/sidekiq/tasks/strategies/rake_task_spec.rb
@@ -21,6 +21,13 @@ RSpec.describe Sidekiq::Tasks::Strategies::RakeTask do
         arg_names: ["bar"]
       )
 
+      expect(File).to receive(:read).with("foo.rb").and_return <<~RUBY
+        namespace :foo do
+          task :bar do
+          end
+        end
+      RUBY
+
       metadata = described_class.new.build_task_metadata(rake_task)
 
       expect(metadata).to be_a(Sidekiq::Tasks::TaskMetadata)
@@ -28,6 +35,28 @@ RSpec.describe Sidekiq::Tasks::Strategies::RakeTask do
       expect(metadata.desc).to eq("Bar")
       expect(metadata.file).to eq("foo.rb")
       expect(metadata.args).to eq(["bar"])
+      expect(metadata.sidekiq_options).to eq({})
+    end
+
+    it "extracts sidekiq_options from a magic comment", :aggregate_failures do
+      rake_task = build_rake_task(
+        name: "foo:bar",
+        full_comment: nil,
+        locations: ["foo.rb:3"],
+        arg_names: []
+      )
+
+      expect(File).to receive(:read).with("foo.rb").and_return <<~RUBY
+        namespace :foo do
+          # sidekiq-tasks:sidekiq_options: queue: critical, retry: 5
+          task :bar do
+          end
+        end
+      RUBY
+
+      metadata = described_class.new.build_task_metadata(rake_task)
+
+      expect(metadata.sidekiq_options).to eq(queue: "critical", retry: 5)
     end
   end